requirements.txt
django==1.5.2
MySQL-python==1.2.4
Pillow==2.1.0
South==0.8.2
django-ckeditor==4.0.2
